feat: light client get cached merkle tree (#1775)
* feat: add Merkle tree getters

refactor: rename RpcConnection -> Rpc

refactor: rename SolanaRpcConnection -> LightClient, rename RpcConnection -> Rpc

rename: SolanaRpcUrl -> RpcUrl

fix: make PhotonIndexer flexible for non local host

feat: add ValidityProofWithContext::pack_merkle_tree_accounts

refactor: pack_merkle_tree_accounts add pack output indices

fix: use TreeType in test-utils

fix: lint

fix: counter example assert

chore: set LightProgramTest address Merkle tree to amt1Ayt45jfbdw5YSo7iz6WZxUmnZsQTYXy82hVwyC2

* refactor: set_account inputs without references

* fix: account compression tests!

* fix: counter example

* rename: indexer::types::Account -> CompressedAccount

* refactor: rename CompressedAccount::merkle_context -> tree_info

* refactor: light-sdk rename merkle_context -> tree_info, use PackedTreeInfo which contains root_index instead of PackedMerkleContext, rename PackedMerkleTreeAccounts -> PackedTreeAccounts and return one output_tree_index

* refactor: split get_compressed_account into two add get_compressed_account_by_hash

* chore: remove duplicate TreeMeta type

* refactor: pack_tree_accounts output

* chore: add log_failed_tx option, log all tx with debug asserts

* refactor: light-sdk rename merkle_context -> tree_info

* fix: light-program-test

* chore: deprecate PackMerkleContext functions

* fix: registry test

* add light-client tests

* test: add program-test indexer trait test

* chore: move light-client test to program-tests

* refactor: rename account_trees -> state_trees, get_output_tree_index -> pack_output_tree_index

* refactor: rename field output_tree_index -> output_state_tree_index

+++ sdk-libs/client/src/indexer/options.rs
@@ -2,10 +2,11 @@ use photon_api::models::{FilterSelector, Memcmp};
 use solana_account_decoder_client_types::UiDataSliceConfig;
 use solana_pubkey::Pubkey;
 
+#[derive(Debug, Clone)]
 pub struct GetCompressedTokenAccountsByOwnerOrDelegateOptions {
     pub mint: Option<Pubkey>,
     pub cursor: Option<String>,
-    pub limit: Option<u64>,
+    pub limit: Option<u16>,
 }
 
 impl GetCompressedTokenAccountsByOwnerOrDelegateOptions {
@@ -23,14 +24,14 @@ impl GetCompressedTokenAccountsByOwnerOrDelegateOptions {
 /// **Limit** is the maximum number of results to return per page.
 pub struct PaginatedOptions {
     pub cursor: Option<String>,
-    pub limit: Option<u64>,
+    pub limit: Option<u16>,
 }
 
 pub struct GetCompressedAccountsByOwnerConfig {
     pub filters: Option<Vec<GetCompressedAccountsFilter>>,
     pub data_slice: Option<UiDataSliceConfig>,
     pub cursor: Option<String>,
-    pub limit: Option<u64>,
+    pub limit: Option<u16>,
 }
 
 #[derive(Clone)]
